Replicate PIN functionality for Passphrase
