+++ shared/enclave/server/views/passphrase.ts
@@ -1,0 +1,8 @@
+import showPrompt from './showPrompt';
+import template from './passphrase.html';
+
+const EVENT = 'enclave:passphrase';
+
+export function showPassphrasePrompt(): Promise<string> {
+  return showPrompt(template, EVENT);
+}
